added functionalities to Schedules and Datepicker
