+++ package.json
@@ -20,6 +20,7 @@
     "bootswatch": "^5.2.3",
     "date-fns": "^2.29.3",
     "dayjs": "^1.11.7",
+    "dhtmlx-scheduler": "^6.0.3",
     "mui-datatables": "^4.3.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
