+++ src/Pages/DatePicker/Datepicker.jsx
@@ -1,0 +1,29 @@
+import * as React from 'react';
+import dayjs from 'dayjs';
+import Box from '@mui/material/Box';
+import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs';
+import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
+import { DatePicker } from '@mui/x-date-pickers/DatePicker';
+
+
+export default function CustomInput() {
+  const [value, setValue] = React.useState(dayjs('2022-04-07'));
+
+  return (
+    <LocalizationProvider dateAdapter={AdapterDayjs}>
+      <DatePicker
+        label="Custom input"
+        value={value}
+        onChange={(newValue) => {
+          setValue(newValue);
+        }}
+        renderInput={({ inputRef, inputProps, InputProps }) => (
+          <Box sx={{ display: 'flex', alignItems: 'center' }}>
+            <input ref={inputRef} {...inputProps} />
+            {InputProps?.endAdornment}
+          </Box>
+        )}
+      />
+    </LocalizationProvider>
+  );
+}
